+++ frontend/src/App/App.css
@@ -1,0 +1,4 @@
+.App {
+  display: flex;
+  height: 100vh;
+}
