+++ frontend/src/App/Calendar/Calendar.css
@@ -1,0 +1,5 @@
+.Calendar {
+  grid-column-start: 1;
+  grid-column-end: 4;
+  background-color: grey;
+}
