+++ frontend/src/App/Lessons/Lessons.css
@@ -1,0 +1,3 @@
+.Lessons {
+  background-color: grey;
+}
