+++ frontend/src/App/Lessons/Lessons.js
@@ -1,0 +1,13 @@
+import React, { Component } from 'react';
+
+import './Lessons.css';
+
+export default class Lessons extends Component {
+  render() {
+    return (
+      <div className="Lessons">
+        { this.props.children }
+      </div>
+    )
+  }
+}
