+++ frontend/src/App/Lessons/Lessons.less
@@ -1,0 +1,3 @@
+.Lessons {
+    background-color: grey;
+}
