+++ frontend/src/App/Menu/Menu.css
@@ -1,0 +1,4 @@
+.Menu {
+  background-color: blueviolet;
+  flex: 0 0 17vw;
+}
