+++ frontend/src/App/Menu/Menu.js
@@ -1,0 +1,13 @@
+import React, { Component } from 'react';
+
+import './Menu.css';
+
+export default class Menu extends Component {
+  render() {
+    return (
+      <div className="Menu">
+        { this.props.children }
+      </div>
+    );
+  }
+}
