+++ frontend/src/App/Menu/Menu.less
@@ -1,0 +1,4 @@
+.Menu {
+    background-color: blueviolet;
+    flex: 0 0 17vw;
+}
