+++ frontend/src/App/Students/Students.css
@@ -1,0 +1,3 @@
+.Students {
+  background-color: grey;
+}
