+++ frontend/src/App/Students/Students.less
@@ -1,0 +1,3 @@
+.Students {
+    background-color: grey;
+}
