+++ frontend/src/App/Today/Today.css
@@ -1,0 +1,3 @@
+.Today {
+  background-color: grey;
+}
