+++ frontend/src/App/Today/Today.less
@@ -1,0 +1,3 @@
+.Today {
+    background-color: grey;
+}
