+++ frontend/src/App/WorkArea/WorkArea.less
@@ -1,0 +1,7 @@
+.WorkArea {
+    flex: 1 1 83vw;
+    display: grid;
+    grid-template-columns: repeat(3, 1fr);
+    grid-gap: 2em;
+    margin: 2em;
+}
