+++ frontend/src/index.css
@@ -1,40 +1,5 @@
 body {
-  margin: 0;
-  padding: 0;
-  font-family: sans-serif;
-}
-
-.App {
-  display: flex;
-  height: 100vh;
-}
-
-.Menu {
-  background-color: blueviolet;
-  flex: 0 0 17vw;
-}
-
-.WorkArea {
-  flex: 1 1 83vw;
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  grid-gap: 2em;
-  margin: 2em;
-}
-
-.Calendar {
-  grid-column: 1/4;
-  background-color: grey;
-}
-
-.Today {
-  background-color: grey;
-}
-
-.Lessons {
-  background-color: grey;
-}
-
-.Students {
-  background-color: grey;
+    margin: 0;
+    padding: 0;
+    font-family: sans-serif;
 }
